requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-login
python-dotenv
flask-socketio
requests
beautifulsoup4
psycopg2-binary
flask_cors
boto3
gunicorn
eventlet
pytest
scikit-learn
google-generativeai